# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Codalla AI Coding Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── Success Criteria ────────────────────────────────────────────────────────
...
  /ai/action:
    post:
      operationId: runCodeAction
      tags: [ai]
      summary: Run a code action (fix, explain, tests, optimize, diff)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeActionInput"
      responses:
        "200":
          description: Action result with optional diff
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodeActionResult"

  # ─── API Keys ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Project:
      type: object
      required: [id, name, localPath, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        isShared:
          type: boolean
        isOwner:
          type: boolean
        localPath:
          type: string
        gitRemoteUrl:
          type: ["string", "null"]
        currentBranch:
          type: string
        lastSynced:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        story:
          type: ["string", "null"]
        target:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
...
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        isShared:
          type: boolean
        gitRemoteUrl:
          type: string
        currentBranch:
          type: string
        story:
          type: string
        target:
          type: string

    # ─── Success Criteria ──────────────────────────────────────────────────────
...
    CodeActionInput:
      type: object
      required: [action, code, modelId, provider]
      properties:
        action:
          type: string
          enum: [fix, explain, tests, optimize, document]
        code:
          type: string
        language:
          type: string
        filename:
          type: string
        modelId:
          type: string
        provider:
          type: string
        context:
          type: string
        conversationId:
          type: string
    CodeActionResult:
      type: object
      required: [action, explanation]
      properties:
        action:
          type: string
        explanation:
          type: string
        suggestedCode:
          type: ["string", "null"]
        diff:
          type: ["string", "null"]
        tokensUsed:
          type: integer
        cost:
          type: number

    # ─── API Keys ──────────────────────────────────────────────────────────────
...