# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Codalla AI Coding Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Project {
  id: string;
  name: string;
  isShared?: boolean;
  isOwner?: boolean;
  localPath: string;
  /** @nullable */
  gitRemoteUrl?: string | null;
  currentBranch?: string;
  /** @nullable */
  lastSynced?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  story?: string | null;
  /** @nullable */
  target?: string | null;
  createdAt: string;
  updatedAt?: string;
}