# lib/api-zod/src/generated/types/projectUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Codalla AI Coding Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectUpdate {
  name?: string;
  description?: string;
  isShared?: boolean;
  gitRemoteUrl?: string;
  currentBranch?: string;
  story?: string;
  target?: string;
}